requirements.txt
pandas
geopandas
matplotlib
seaborn
plotly
numpy
folium
scikit-learn
shapely
mapclassify
openpyxl
langchain
langgraph
langchain-openai
langchain-community
chromadb
langchain-chroma
gradio
fastapi
uvicorn
python-dotenv
tiktoken
sentence-transformers
langchain-huggingface
osmnx
osmnx
duckduckgo-search
